safeguard and refactor internals initialization logic
