+++ src/canvas/main.tsx
@@ -4,7 +4,7 @@ import { WidgetInternal } from "../types";
 import App from "./App";
 import { invoke } from "@tauri-apps/api";
 
-invoke<Record<string, WidgetInternal>>("get_widget_internals")
+invoke<Record<string, WidgetInternal>>("init_widget_internals")
   .then((widgetInternals) => {
     createRoot(document.getElementById("canvas")!).render(
       <StrictMode>
